refactor: ModItemDetailView.swift
