+++ BaldursModManager/Utilities/LsxUtility.swift
@@ -1,0 +1,11 @@
+//
+//  LsxUtility.swift
+//  BaldursModManager
+//
+//  Created by Justin Bush on 1/7/24.
+//
+
+import Foundation
+
+struct LsxUtility {
+}
